+++ src/id_translation/fetching/__init__.py
@@ -17,7 +17,6 @@ Base fetchers:
 """
 
 from ._abstract_fetcher import AbstractFetcher
-from ._cache import CacheAccess, CacheMetadata
 from ._fetcher import Fetcher
 from ._memory_fetcher import MemoryFetcher
 from ._multi_fetcher import MultiFetcher
@@ -27,8 +26,6 @@ from ._sql_fetcher import SqlFetcher
 __all__ = [
     "Fetcher",
     "AbstractFetcher",
-    "CacheAccess",
-    "CacheMetadata",
     "MemoryFetcher",
     "MultiFetcher",
     "PandasFetcher",